requirements.txt
python-telegram-bot==21.6
psycopg2-binary==2.9.9
gspread==6.1.2
google-auth==2.29.0
streamlit==1.33.0
pandas==2.2.2
